CA-315962 cleanupwatchdog: Allow resetting watchdog slots without an HA file
In some cases the slots are used but xhad was not enable to write to
file to keep track of them.
This behaviour allows xhad to start up in these cases without needed to
reboot the host.
The program exits when the system does not have enough memory or the
hypercall returned an unexpected result but not when the slot id was not
enabled yet.

Signed-off-by: Pau Ruiz Safont <pau.safont@citrix.com>