CI: add CodeChecker (Clang-tidy + Clang Static Analyzer)
Convert to SARIF format and upload, which will show up on the Security tab under 'Code Scanning'.

The HTML report is available as an artifact too.

GCC would also support static analysis and SARIF output, but github can't always
process the output from it, so for now use just the clang one.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>