CA-407106: Fix potential uninitialized value during timeout
If a timeout is reached, then the loop is never entered, and 'index' is uninitialized.
All the logging calls would use an uninitialized value, and later `my_sfdomain` would also be uninitialized
and various decisions would be taken based on its state.

Ensure that these variables are initialized at least once in the loop by moving the timeout condition.

It is very unlikely that this'd actually happen, but without this change the static analyzer
wouldn't be able to spot other bugs that'd leave these values uninitialized.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>