CA-407106: Avoid using deprecated sigset and sigignore functions
Use `sigaction` and `sigprocmask` instead.
There is also `signal`, but the manpage says to not use it.
Indeed `sigaction` has more reliable semantics by default in POSIX (e.g. it blocks the signal while you are handling it,
so you don't reenter it while it is executing).

Unfortunately the new functions are more verbose, so introduce a loop for setting up the signal handlers.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>