CA-407107: drop release mode, it is unused
xha.spec does a 'make install' which installs the debug build by default.
The only difference between debug and release is NDEBUG (no -O flags are used),
but considering that we've always shipped the debug version in production builds,
keep that one.

There are some advantages to shipping the debug version:
* stacktraces may work better than at O2
* if the assertions finds a bug in the code it might be better to stop, 
  rather than continue and do something incorrect (corrupt memory, etc.).
  Given that in the past years that we've been shipping XHA with asserts on we haven't really found
  failed assertions, keep them on.

Signed-off-by: Edwin Török <edwin.torok@cloud.com>